realtime-transcriptions/requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-speech==0.36.3
googleapis-common-protos==1.5.8
grpcio==1.19.0
idna==2.8
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
websockets==7.0
